updated script to latest changes

+++ deploy_dev.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 
+USER=$1
+
+if [ -z "$USER" ]; then
+    echo "please enter a user to deploy for (hugo or samuel)"
+    exit 1
+fi
+
 rm -fr dist
 npm run build
-ssh my24-dev "rm -fr ~/frontend/*"
-scp -r dist/* my24-dev:~/frontend
+ssh $USER@my24-dev "rm -fr /var/www/developers/$USER/frontend/*"
+scp -r dist/* $USER@my24-dev:/var/www/developers/$USER/frontend/
 rm -fr dist
